Moved ctags test into ctags package
