+++ src/arduino.cc/builder/ctags/test_data/TestCTagsParserDefaultArguments.txt
