+++ src/arduino.cc/builder/ctags/test_data/TestCTagsParserShouldDealFunctionWithDifferentSignatures.txt
