+++ src/arduino.cc/builder/ctags/test_data/TestCTagsParserShouldDealWithClasses.txt
