+++ src/arduino.cc/builder/ctags/test_data/TestCTagsParserShouldDealWithMacros.txt
