+++ src/arduino.cc/builder/ctags/test_data/TestCTagsParserShouldListTemplates2.txt
