+++ src/arduino.cc/builder/ctags/test_data/TestCTagsRunnerSketchWithClassFunction.txt
